Revert "Merge pull request #12637 from ichard26/refactor/imports"
This reverts commit 71a08a71bc732f6f8aa48814f731e9bcfafeafb6, reversing
changes made to 22142d6fcc9885587bcf102d7652529ec1d9503e.

This patch made the pip self-version check import
`pip._internal.self_outdated_check` on demand. This blows up when
downgrading (and possibly upgrading) pip as it imports the module
_after_ the downgrade, which is problematic as the older self-check logic
is incompatible with the rest of the pip already loaded into memory.
